+++ Engine/src/Renderer/Scene.cpp
@@ -14,7 +14,7 @@ namespace SceneEditor{
 		}
 	}
 
-	std::shared_ptr<ModelController> Scene::AddModel(const std::string_view& Path, uint32_t& NoVertices)
+	std::shared_ptr<ModelController> Scene::AddModel(const std::string_view& Path, uint64_t& NoVertices)
 	{
 		auto&& model = std::make_shared<Model>(Path.data());
 		m_SceneModels.emplace_back(model);
