+++ Engine/src/Renderer/Scene.h
@@ -20,7 +20,7 @@ namespace SceneEditor{
 		void AddPass(std::unique_ptr<Pass>& Pass_p);
 		void ClearScene( ) { m_SceneModels.clear( ); };
 
-		std::shared_ptr<ModelController> AddModel(const std::string_view& Path, uint32_t& NoVertices);
+		std::shared_ptr<ModelController> AddModel(const std::string_view& Path, uint64_t& NoVertices);
 		std::shared_ptr<ModelController> AddModel(std::shared_ptr<Model>& Model);
 
 		void DeleteModel(uint32_t Index) { m_SceneModels.erase(m_SceneModels.begin() + Index); };
